+++ pybop/__init__.py
@@ -96,7 +96,7 @@ from ._experiment import Experiment
 #
 # Optimiser class
 #
-from ._optimisation import Optimisation
+from .optimisers._optimisation import BaseOptimiser, Optimisation
 from .optimisers.base_optimiser import BasePintsOptimiser
 from .optimisers.scipy_optimisers import (
     BaseSciPyOptimiser,